// src/app/models/category.model.ts
export enum Category {
  BAKERY = "Bakery",
  FRUIT_AND_VEGETABLES = "Fruit and vegetables",
  MEAT_AND_FISH = "Meat and fish",
  DRINKS = "Drinks",
  KITCHEN = "Kitchen",
  SPECIAL_NUTRITION = "Special nutrition",
  BABY = "Baby",
  PHARMACY = "Pharmacy",
}

export enum SubCategory {
  VEGETABLES = "Vegetables",
  FRUIT = "Fruit",
  MEAT = "Meat",
  FISH = "Fish",
  BAKERY = "Bakery",
  SPECIAL_NUTRITION = "Special nutrition",
  DRINKS = "Drinks",
  KITCHEN = "Kitchen",
}

export type ListItem = {
  name: string
}